Allow for automation by accepting a refresh_token param
If the refresh token is provided, new OAuth access and ID tokens can be
requested without user intervention. This allows for automation.

To facilitate this, offline access is requested and the refresh token is
returned in the shell format responses.

Another tool can use these two pieces to automate the fetching of fresh
AWS keys and tokens each day.